Refactor `Container` struct to store its process state
This commit also refactors how status codes are transmitted. Ultimately,
this enables wait requests on stopped containers to respond with their
exit codes.

Old behavior:
`TaskService` stores `wait_channels` and adds channels to it when `wait`
is called. `handle_signals` sends status codes through the channels in
`wait_channels` when a container exits.

New behavior:
`Container` stores `wait_channels` and exposes a function to create new
channels and exposes a function to transmit to all channels.
`handle_signals` sends status codes through a new, independent channel.
When status codes are received on this channel, `Container` state is
updated and the function to transmit to all channels is called.